requirements.txt
mido>=1.3.0
python-rtmidi>=1.5.6